Fix compatibility issue with game version 1.0.16.0.22

+++ ToolbeltFix/Patches/HotkeyController_Patches.cs
@@ -697,7 +697,7 @@ namespace ToolbeltFix.Patches
                     SlotStorage storage = _playerRef(__instance).Inventory.GetSlotStorage() as SlotStorage;
 
                     JObject jobject = new JObject();
-                    jobject.AddField("Level", _levelRef(__instance));
+                    jobject.AddField("Level", (float)_levelRef(__instance));
                     JObject jobject2 = new JObject();
                     for (int i = 0; i < _hotkeysRef(__instance).Length; i++)
                     {
